Refactor producer metrics
