+++ app/src/main/java/org/astraea/app/cost/NodeThroughputCost.java
@@ -23,9 +23,9 @@ import java.util.Optional;
 import java.util.stream.Collectors;
 import org.astraea.app.admin.ClusterBean;
 import org.astraea.app.admin.ClusterInfo;
-import org.astraea.app.metrics.KafkaMetrics;
 import org.astraea.app.metrics.collector.Fetcher;
 import org.astraea.app.metrics.producer.HasProducerNodeMetrics;
+import org.astraea.app.metrics.producer.ProducerMetrics;
 
 public class NodeThroughputCost implements HasBrokerCost {
 
@@ -56,6 +56,6 @@ public class NodeThroughputCost implements HasBrokerCost {
 
   @Override
   public Optional<Fetcher> fetcher() {
-    return Optional.of(KafkaMetrics.Producer::nodes);
+    return Optional.of(ProducerMetrics::nodes);
   }
 }
