+++ app/src/main/java/org/astraea/app/metrics/KafkaMetrics.java
@@ -18,12 +18,9 @@ package org.astraea.app.metrics;
 
 import java.util.Arrays;
 import java.util.Collection;
-import java.util.Map;
 import java.util.stream.Collectors;
 import org.astraea.app.metrics.broker.HasValue;
 import org.astraea.app.metrics.broker.TotalTimeMs;
-import org.astraea.app.metrics.producer.HasProducerNodeMetrics;
-import org.astraea.app.metrics.producer.HasProducerTopicMetrics;
 
 public final class KafkaMetrics {
 
@@ -82,80 +79,6 @@ public final class KafkaMetrics {
           .stream()
           .map(HasValue::of)
           .collect(Collectors.toUnmodifiableList());
-    }
-  }
-
-  public static final class Producer {
-
-    private static int brokerId(String node) {
-      return Integer.parseInt(node.substring(node.indexOf("-") + 1));
-    }
-
-    private Producer() {}
-
-    /**
-     * node metrics traced by producer
-     *
-     * @param mBeanClient to query beans
-     * @param brokerId broker ids
-     * @return key is client id used by producer, and value is node metrics traced by each producer
-     */
-    public static Map<String, HasProducerNodeMetrics> node(MBeanClient mBeanClient, int brokerId) {
-      return mBeanClient
-          .queryBeans(
-              BeanQuery.builder()
-                  .domainName("kafka.producer")
-                  .property("type", "producer-node-metrics")
-                  .property("node-id", "node-" + brokerId)
-                  .property("client-id", "*")
-                  .build())
-          .stream()
-          .collect(
-              Collectors.toUnmodifiableMap(
-                  b -> b.properties().get("client-id"),
-                  b -> HasProducerNodeMetrics.of(b, brokerId(b.properties().get("node-id")))));
-    }
-
-    /**
-     * collect HasProducerNodeMetrics from all producers.
-     *
-     * @param mBeanClient to query metrics
-     * @return key is broker id, and value is associated to broker metrics recorded by all producers
-     */
-    public static Collection<HasProducerNodeMetrics> nodes(MBeanClient mBeanClient) {
-      return mBeanClient
-          .queryBeans(
-              BeanQuery.builder()
-                  .domainName("kafka.producer")
-                  .property("type", "producer-node-metrics")
-                  .property("node-id", "*")
-                  .property("client-id", "*")
-                  .build())
-          .stream()
-          .map(b -> HasProducerNodeMetrics.of(b, brokerId(b.properties().get("node-id"))))
-          .collect(Collectors.toUnmodifiableList());
-    }
-
-    /**
-     * topic metrics traced by producer
-     *
-     * @param mBeanClient to query beans
-     * @param topic topic name
-     * @return key is client id used by producer, and value is topic metrics traced by each producer
-     */
-    public static Map<String, HasProducerTopicMetrics> topic(
-        MBeanClient mBeanClient, String topic) {
-      return mBeanClient
-          .queryBeans(
-              BeanQuery.builder()
-                  .domainName("kafka.producer")
-                  .property("type", "producer-topic-metrics")
-                  .property("client-id", "*")
-                  .property("topic", topic)
-                  .build())
-          .stream()
-          .collect(
-              Collectors.toUnmodifiableMap(b -> b.properties().get("client-id"), b -> () -> b));
     }
   }
 }
